server/requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
PyJWT==0.4.3
SQLAlchemy==0.9.8
Werkzeug==0.10.1
argparse==1.2.1
gunicorn==19.2.1
itsdangerous==0.24
psycopg2==2.6
wsgiref==0.1.2
